Cristal-Ollama-master/pyproject.toml
[tool.poetry]
name = "cristal-ollama"
version = "0.1.0"
description = ""
authors = ["GuiMcs00 <guilherme.amaral2004@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pymongo = "^4.7.3"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
pytest = "^8.2.2"
ollama = "^0.2.0"
numpy = "^1.26.4"
faiss-cpu = "^1.8.0"
transformers = {extras = ["torch"], version = "^4.41.2"}
jupyter = "^1.0.0"
torch = "^2.3.0"
pandas = "^2.2.2"
scikit-learn = "^1.5.0"
sentencepiece = "^0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

include = [
    "src/**",
    "src/DataBusiness/**",
    "src/FineTuning/**",
    "src/FineTuning/config/**",
    "src/FineTuning/dataset/**",
    "src/FineTuning/models/**",
    "src/FineTuning/scripts/**",
    "src/FineTuning/utils/**"
]